+++ Spring.pde
@@ -5,13 +5,13 @@ public class Spring<Atom> extends Edge implements Renderable {
 
         public void render() {
                 pushMatrix();
-                       stroke(Constants.DEFAULT_SPRING_R,
-                              Constants.DEFAULT_SPRING_G,
-                              Constants.DEFAULT_SPRING_B,
-                              Constants.DEFAULT_SPRING_A);
-                       strokeWeight(Constants.DEFAULT_SPRING_STROKE_WEIGHT); 
-                       line(((Node)vertex1).getX(), ((Node)vertex1).getY(), ((Node)vertex1).getZ(),
-                            ((Node)vertex2).getX(), ((Node)vertex2).getY(), ((Node)vertex2).getZ());
+                       stroke(Constants.DEFAULT_SPRING_R.floatValue(),
+                              Constants.DEFAULT_SPRING_G.floatValue(),
+                              Constants.DEFAULT_SPRING_B.floatValue(),
+                              Constants.DEFAULT_SPRING_A.floatValue());
+                       strokeWeight(Constants.DEFAULT_SPRING_STROKE_WEIGHT.floatValue()); 
+                       line(((Node)vertex1).getX().floatValue(), ((Node)vertex1).getY().floatValue(), ((Node)vertex1).getZ().floatValue(),
+                            ((Node)vertex2).getX().floatValue(), ((Node)vertex2).getY().floatValue(), ((Node)vertex2).getZ().floatValue());
                popMatrix();
         }
 }
