+++ Utilities.java
@@ -1,24 +1,24 @@
 public class Utilities {
 
-        public static final Float EPSILON = 0.000000000001f;
-        public static final Float UNIT = 1f;
+        public static final Double EPSILON = 0.000000000001;
+        public static final Double UNIT = 1.0;
 
-        public static Float abs(Float a) {
+        public static Double abs(Double a) {
                 if (a > EPSILON) {
                         return a;
                 } else if (a < EPSILON) {
                         return -1 * a;
                 } else {
-                        return 0f;
+                        return 0.0;
                 }
         }
 
-        public static Boolean inRange(Float a, Float b) {
+        public static Boolean inRange(Double a, Double b) {
                 return abs(a - b) < EPSILON;
         }
 
-        public static Boolean inUnit(Float a) {
-                Float val = abs(a);
+        public static Boolean inUnit(Double a) {
+                Double val = abs(a);
                 return (val < UNIT) || (inRange(val, UNIT));
         }
